Moved trajectory generator file into new folder and made refactoring changes
